requirements.txt
streamlit
opencv-python
google-generativeai
pillow
numpy
